chore: Remove unused kafka-deduplicator code
