+++ rust/kafka-deduplicator/src/kafka/batch_context.rs
@@ -2,11 +2,20 @@ use std::sync::Arc;
 use tokio::runtime::Handle;
 use tokio::sync::mpsc;
 
+use crate::kafka::rebalance_handler::RebalanceHandler;
 use crate::kafka::types::Partition;
-use crate::kafka::{rebalance_handler::RebalanceHandler, stateful_context::RebalanceEvent};
 use rdkafka::consumer::{BaseConsumer, ConsumerContext, Rebalance};
 use rdkafka::{ClientContext, TopicPartitionList};
 use tracing::{error, info, warn};
+
+/// Events sent to the async rebalance worker
+#[derive(Debug, Clone)]
+pub enum RebalanceEvent {
+    /// Partitions are being revoked
+    Revoke(Vec<Partition>),
+    /// Partitions have been assigned
+    Assign(Vec<Partition>),
+}
 
 pub struct BatchConsumerContext {
     rebalance_handler: Arc<dyn RebalanceHandler>,
