+++ rust/kafka-deduplicator/src/kafka/test_utils.rs
@@ -1,37 +1,15 @@
 use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::{Arc, Mutex};
+use std::sync::Mutex;
 
 use async_trait::async_trait;
 
 use super::rebalance_handler::RebalanceHandler;
-use super::stateful_context::StatefulConsumerContext;
-use super::tracker::InFlightTracker;
 use super::types::Partition;
 
 use anyhow::Result;
-use rdkafka::config::ClientConfig;
-use rdkafka::consumer::BaseConsumer;
 use rdkafka::TopicPartitionList;
 
-/// Test utilities for kafka module tests
-pub fn create_test_consumer<H: RebalanceHandler + 'static>(
-    handler: Arc<H>,
-) -> BaseConsumer<StatefulConsumerContext> {
-    let tracker = Arc::new(InFlightTracker::new());
-    let context = StatefulConsumerContext::new(handler, tracker);
-
-    let consumer: BaseConsumer<StatefulConsumerContext> = ClientConfig::new()
-        .set("bootstrap.servers", "localhost:9092")
-        .set("group.id", "test-group")
-        .set("enable.auto.commit", "false")
-        .set("auto.offset.reset", "earliest")
-        .create_with_context(context)
-        .expect("Consumer creation failed");
-
-    consumer
-}
-
-// Test implementation of RebalanceHandler that tracks calls
+/// Test implementation of RebalanceHandler that tracks calls
 #[derive(Default)]
 pub struct TestRebalanceHandler {
     pub assigned_count: AtomicUsize,
