+++ rust/kafka-deduplicator/src/lib.rs
@@ -2,14 +2,12 @@ pub mod checkpoint;
 pub mod checkpoint_manager;
 pub mod config;
 pub mod deduplication_batch_processor;
-pub mod deduplication_processor;
 pub mod duplicate_event;
 pub mod duplicate_metrics;
 pub mod event;
 pub mod kafka;
 pub mod metrics;
 pub mod metrics_const;
-pub mod processor_pool;
 pub mod processor_rebalance_handler;
 pub mod rocksdb;
 pub mod service;
